requirements.txt
# Backend API
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0

# Data Science & ML
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.7.2
matplotlib==3.10.7
scipy==1.16.3

# MLOps Stack
mlflow==3.6.0
apache-airflow==3.1.3

# Frontend
streamlit==1.31.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
